DRF/blogAPI/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==4.2.15
django-allauth==64.0.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
